Address review: fold via transform_point, single clamp, TwoPointRadial naming, shared test helpers
- fold_linear_gradient_transform maps the endpoints with
  Transform2D::transform_point; only the A^-T direction math reads the
  matrix components.
- radialTwoPointT sets t unclamped per branch and clamps once before
  returning (WGSL now matches the GLSL); the discriminant is immutable
  with sqrt(max(disc, 0.0)), and the containment case folds into the
  branch condition so a negative discriminant only counts as a miss for
  a > 0. Both backends.
- ShaderType::FillGradientRadial / FillImageGradientRadial become
  FillGradientTwoPointRadial / FillImageGradientTwoPointRadial, with the
  shader functions renamed to match, so they cannot be read as the conic
  (angular) gradient shaders.
- The headless device and render/readback helpers move to
  tests/common/mod.rs and the conformance test uses them.
